LayoutTests/imported/w3c/web-platform-tests/html/dom/historical-expected.txt


PASS document.applets is always empty
PASS <attachment> is HTMLUnknownElement
PASS <applet> is HTMLUnknownElement
PASS <layer> is HTMLUnknownElement
PASS <nolayer> is HTMLUnknownElement
PASS HTMLAppletElement is no more
PASS document.all cannot find applet
PASS document cannot find applet
PASS window cannot find applet
PASS applet is not styled
PASS HTMLTableDataCellElement interface is removed
PASS HTMLTableHeaderCellElement interface is removed
